+++ Miguel/Exercicios/Capitulo_2/ex4_consumo_combustivel/middleware.h
@@ -3,10 +3,10 @@
 
 void clear_buffer();
 
-int ask_kilometers_traveled();
+float ask_kilometers_traveled();
 
-int ask_gasoline_consumed();
+float ask_gasoline_consumed();
 
-float calculate_consumption(int kilometers_traveled, int gasoline_consumed);
+float calculate_consumption(float kilometers_traveled, float gasoline_consumed);
 
 void show_consumption(float average_consumption);
